Pull Feature up out of Features
This class needs a test, and this simplifies a couple of specs
